Programmaticly cut syllable sounds
Converts the playback logic to use untouched clips from the sound bank
with separately defined offsets (in seconds) when forming sounds.

The initial start and end offsets are more aggressive compared to the
original manually cut files, but definitely need improvement still.

This is essentially most of the work towards #30.